+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.4](https://github.com/OxideAV/oxideav-pcx/compare/v0.0.3...v0.0.4) - 2026-07-18
+## [0.1.0](https://github.com/OxideAV/oxideav-pcx/compare/v0.0.3...v0.1.0) - 2026-07-18
 
 ### Other